requirements.txt
schedule==1.2.0
pyautogui==0.9.54
pygetwindow==0.0.9
pandas==2.0.3
psycopg2-binary==2.9.9
plotly==5.17.0
streamlit==1.28.1